+++ src/presentation/controller/signup.spec.ts
@@ -1,5 +1,5 @@
 import { SignUpController } from './signup';
-import { EmailValidator } from '../protocols/email-validator.interface';
+import { EmailValidator } from '../protocols';
 import { MissingParamError, InvalidParamError, ServerError } from '../errors';
 
 interface SutTypes {
