+++ src/presentation/protocols/index.ts
@@ -1,0 +1,3 @@
+export * from './controller.interface';
+export * from './email-validator.interface';
+export * from './http.interface';
